style: fix cosmetic clippy lints, re-enable -D warnings in CI
cargo clippy --fix did the bulk of the cleanup; remaining hand-fixes:

butteraugli-gpu:
- src/kernels/blur.rs: replace LOG2_E literal with std::f32::consts::LOG2_E.
- src/pipeline_batch.rs: drop double parens around (w*h) in two reduce_*
  calls.

zensim-gpu:
- src/kernels/fused.rs: replace 'let mut s_val = 0.0; if … { s_val = …; }'
  channel switch with 'let s_val = if … else if … else …;' (no
  needless_late_init, no unused_assignment); drop dead TX_US constant;
  add #![allow(clippy::doc_overindented_list_items)] for the algorithm
  docstring (clippy's suggested 3-space indent would visually misalign
  the sub-list continuations).
- src/kernels/reduce.rs: 'sum = sum + x' → 'sum += x'.
- src/pipeline.rs: replace two manual-memcpy loops with copy_from_slice;
  add #[allow(dead_code)] to Scale (transitional fields kept around for
  the in-progress per-channel debug dump).

ssim2-gpu:
- src/kernels/blur.rs, src/pipeline_batch.rs: blank-line separators
  before / after numbered lists in module docstrings to satisfy
  doc_lazy_continuation.

zen-metrics-cli:
- src/decode.rs: #[allow(dead_code)] on Rgb8Image::pixels (it's used by
  every metric backend; only flagged in feature combos that disable
  all backends).

CI:
- ci.yml: re-enabled -D warnings, dropped the advisory mode workaround.

Verified: 0 clippy warnings on workspace; 45/45 tests pass on CUDA.

+++ crates/butteraugli-gpu/src/kernels/reduction.rs
@@ -272,8 +272,8 @@ mod portable {
         threads_per_image: u32,
     ) {
         let batch_idx = CUBE_POS_Y;
-        let tid_in_plane = UNIT_POS_X + CUBE_POS_X * (CUBE_DIM_X as u32);
-        let stride_per_plane = CUBE_COUNT_X as u32 * (CUBE_DIM_X as u32);
+        let tid_in_plane = UNIT_POS_X + CUBE_POS_X * CUBE_DIM_X;
+        let stride_per_plane = CUBE_COUNT_X * CUBE_DIM_X;
         let plane_off = (batch_idx * plane_stride) as usize;
         let plane_us = plane_stride as usize;
 
@@ -504,8 +504,8 @@ fn batched_max_reduce_kernel(
     plane_stride: u32,
 ) {
     let batch_idx = CUBE_POS_Y;
-    let tid_in_plane = UNIT_POS_X + CUBE_POS_X * (CUBE_DIM_X as u32);
-    let stride_per_plane = CUBE_COUNT_X as u32 * (CUBE_DIM_X as u32);
+    let tid_in_plane = UNIT_POS_X + CUBE_POS_X * CUBE_DIM_X;
+    let stride_per_plane = CUBE_COUNT_X * CUBE_DIM_X;
     let plane_off = (batch_idx * plane_stride) as usize;
     let plane_us = plane_stride as usize;
